feat [SRTP-724] refactoring
